Move contents of Pkg2Types into Types
