+++ src/Pkg2/reqs.jl
@@ -6,7 +6,6 @@ import Base: ==
 import ..PkgError
 using ..Reqs
 using ...Types
-using ...Types.Pkg2Types
 
 # representing lines of REQUIRE files
