+++ src/Pkg2/resolve.jl
@@ -7,7 +7,6 @@ include(joinpath("resolve", "interface.jl"))
 include(joinpath("resolve", "maxsum.jl"))
 
 using ...Types
-using ...Types.Pkg2Types
 using ..Query, .PkgToMaxSumInterface, .MaxSum
 import ..PkgError
