+++ test/resolve.jl
@@ -4,7 +4,6 @@ module ResolveTest
 
 using ..Test
 using Pkg3.Types
-using Pkg3.Types.Pkg2Types
 using Pkg3.Pkg2.Query
 using Pkg3.Pkg2.Resolve
 using Pkg3.Pkg2.Resolve.VersionWeights
